moving entity link API to a specific server lib
