+++ client/src/modules/journal/templates/references.link.html
@@ -2,7 +2,7 @@
 <div class="ui-grid-cell-contents">
   <a 
     style="text-align:center !important"
-    ng-href="journal/reference_transaction/{{row.entity.hrReference}}/{{grid.appScope.languages.key}}"
+    ng-href="entityLink/{{row.entity.hrReference}}/{{grid.appScope.languages.key}}"
     target="_blank"
     ng-show="row.entity.hrReference!=null"  
   >
